+++ exporter/tencentcloudlogserviceexporter/logsdata_to_logservice.go
@@ -33,7 +33,6 @@ const (
 	clsLogTimeUnixNano   = "timeUnixNano"
 	clsLogSeverityNumber = "severityNumber"
 	clsLogSeverityText   = "severityText"
-	clsLogName           = "name"
 	clsLogContent        = "content"
 	clsLogAttribute      = "attribute"
 	clsLogFlags          = "flags"
@@ -57,7 +56,7 @@ func convertLogs(ld pdata.Logs) []*cls.Log {
 		for j := 0; j < ills.Len(); j++ {
 			ils := ills.At(j)
 			instrumentationLibraryContents := instrumentationLibraryToLogContents(ils.InstrumentationLibrary())
-			logs := ils.Logs()
+			logs := ils.LogRecords()
 			for j := 0; j < logs.Len(); j++ {
 				clsLog := mapLogRecordToLogService(logs.At(j), resourceContents, instrumentationLibraryContents)
 				if clsLog != nil {
@@ -158,10 +157,6 @@ func mapLogRecordToLogService(lr pdata.LogRecord,
 		{
 			Key:   proto.String(clsLogSeverityText),
 			Value: proto.String(lr.SeverityText()),
-		},
-		{
-			Key:   proto.String(clsLogName),
-			Value: proto.String(lr.Name()),
 		},
 		{
 			Key:   proto.String(clsLogAttribute),