+++ processor/groupbyattrsprocessor/config_test.go
@@ -15,7 +15,7 @@
 package groupbyattrsprocessor
 
 import (
-	"path"
+	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -36,7 +36,7 @@ func TestLoadConfig(t *testing.T) {
 	err = configtest.CheckConfigStruct(factory.CreateDefaultConfig())
 	require.NoError(t, err)
 
-	cfg, err := servicetest.LoadConfigAndValidate(path.Join(".", "testdata", "config.yaml"), factories)
+	cfg, err := servicetest.LoadConfigAndValidate(filepath.Join("testdata", "config.yaml"), factories)
 
 	require.Nil(t, err)
 	require.NotNil(t, cfg)