+++ receiver/hostmetricsreceiver/internal/scraper/memoryscraper/memory_scraper_test.go
@@ -23,6 +23,7 @@ import (
 	"github.com/shirou/gopsutil/v3/mem"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/model/pdata"
 	"go.opentelemetry.io/collector/receiver/scrapererror"
 
@@ -35,27 +36,53 @@ func TestScrape(t *testing.T) {
 		name              string
 		virtualMemoryFunc func() (*mem.VirtualMemoryStat, error)
 		expectedErr       string
+		initializationErr string
+		config            *Config
+		bootTimeFunc      func() (uint64, error)
 	}
 
 	testCases := []testCase{
 		{
 			name: "Standard",
+			config: &Config{
+				Metrics: metadata.DefaultMetricsSettings(),
+			},
 		},
 		{
 			name:              "Error",
 			virtualMemoryFunc: func() (*mem.VirtualMemoryStat, error) { return nil, errors.New("err1") },
 			expectedErr:       "err1",
+			config: &Config{
+				Metrics: metadata.DefaultMetricsSettings(),
+			},
+		},
+		{
+			name:              "Error",
+			bootTimeFunc:      func() (uint64, error) { return 100, errors.New("err1") },
+			initializationErr: "err1",
+			config: &Config{
+				Metrics: metadata.DefaultMetricsSettings(),
+			},
 		},
 	}
 
 	for _, test := range testCases {
 		t.Run(test.name, func(t *testing.T) {
-			scraper := newMemoryScraper(context.Background(), &Config{})
+			scraper := newMemoryScraper(context.Background(), test.config)
 			if test.virtualMemoryFunc != nil {
 				scraper.virtualMemory = test.virtualMemoryFunc
 			}
+			if test.bootTimeFunc != nil {
+				scraper.bootTime = test.bootTimeFunc
+			}
 
-			md, err := scraper.Scrape(context.Background())
+			err := scraper.start(context.Background(), componenttest.NewNopHost())
+			if test.initializationErr != "" {
+				assert.EqualError(t, err, test.initializationErr)
+				return
+			}
+			require.NoError(t, err, "Failed to initialize memory scraper: %v", err)
+			md, err := scraper.scrape(context.Background())
 			if test.expectedErr != "" {
 				assert.EqualError(t, err, test.expectedErr)
 
@@ -72,7 +99,7 @@ func TestScrape(t *testing.T) {
 			assert.Equal(t, 1, md.MetricCount())
 
 			metrics := md.ResourceMetrics().At(0).InstrumentationLibraryMetrics().At(0).Metrics()
-			assertMemoryUsageMetricValid(t, metrics.At(0), metadata.Metrics.SystemMemoryUsage.New())
+			assertMemoryUsageMetricValid(t, metrics.At(0), "system.memory.usage")
 
 			if runtime.GOOS == "linux" {
 				assertMemoryUsageMetricHasLinuxSpecificStateLabels(t, metrics.At(0))
@@ -85,8 +112,8 @@ func TestScrape(t *testing.T) {
 	}
 }
 
-func assertMemoryUsageMetricValid(t *testing.T, metric pdata.Metric, descriptor pdata.Metric) {
-	internal.AssertDescriptorEqual(t, descriptor, metric)
+func assertMemoryUsageMetricValid(t *testing.T, metric pdata.Metric, expectedName string) {
+	assert.Equal(t, expectedName, metric.Name())
 	assert.GreaterOrEqual(t, metric.Sum().DataPoints().Len(), 2)
 	internal.AssertSumMetricHasAttributeValue(t, metric, 0, metadata.Attributes.State, pdata.NewAttributeValueString(metadata.AttributeState.Used))
 	internal.AssertSumMetricHasAttributeValue(t, metric, 1, metadata.Attributes.State, pdata.NewAttributeValueString(metadata.AttributeState.Free))