+++ receiver/postgresqlreceiver/factory.go
@@ -14,8 +14,6 @@
 
 package postgresqlreceiver // import "github.com/open-telemetry/opentelemetry-collector-contrib/receiver/postgresqlreceiver"
 
-//go:generate mdatagen metadata.yaml
-
 import (
 	"context"
 	"time"
@@ -27,6 +25,8 @@ import (
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/receiver/receiverhelper"
 	"go.opentelemetry.io/collector/receiver/scraperhelper"
+
+	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/postgresqlreceiver/internal/metadata"
 )
 
 const (
@@ -55,6 +55,7 @@ func createDefaultConfig() config.Receiver {
 			InsecureSkipVerify: true,
 		},
 		Databases: make([]string, 0),
+		Metrics:   metadata.DefaultMetricsSettings(),
 	}
 }