+++ CHANGELOG.md
@@ -4,6 +4,40 @@
 
 ## 💡 Enhancements 💡
 
+- `hostreceiver/memoryscraper`: Migrate the scraper to the mdatagen metrics builder (#7312)
+- `lokiexporter`: Use record attributes as log labels (#7569)
+- `routingprocessor`: Do not err on failure to build exporters (#7423)
+- `apachereceiver`: Update to mdatagen v2 (#7573)
+- `datadogexporter`: Don't send host metadata if hostname is empty (#7426)
+- `datadogexporter`: Add insecure_skip_verify flag to configuration (#7422)
+
+## 🛑 Breaking changes 🛑
+
+- `apachereceiver`: Update instrumentation library name from `otel/apache` to `otelcol/apache` (#7754)
+- `prometheusreceiver`: The OpenCensus-based metric conversion pipeline has 
+  been removed.
+  -  The `receiver.prometheus.OTLPDirect` feature gate has been removed as 
+     the direct pipeline is the only remaining pipeline.
+
+## 🚩 Deprecations 🚩
+
+- Deprecated log_names setting from filter processor. (#7552)
+
+## 🧰 Bug fixes 🧰
+
+ - `tailsamplingprocessor`: "And" policy only works as a sub policy under a composite policy (#7590) 
+ - `prometheusreceiver`: Correctly map description and units when converting
+  Prometheus metadata directly to pdata. (#7748)
+ - `sumologicexporter`: fix exporter panics on malformed histogram (#7548)
+
+## 🚀 New components 🚀
+
+## v0.44.0
+
+## 💡 Enhancements 💡
+
+- `dynatraceexporter`: Write error logs using plugin logger (#7360)
+- `dynatraceexporter`: Fix docs for TLS settings (#7568)
 - `tanzuobservabilityexporter`: Turn on metrics exporter (#7281)
 - `attributesprocessor` `resourceprocessor`: Add `from_context` value source
 - `resourcedetectionprocessor`: check cluster config to verify resource is on aws for eks resources (#7186)
@@ -13,9 +47,24 @@
 - `awscontainerinsightreceiver`: add full pod name when configured to AWS Container Insights Receiver (#7415)
 - `hostreceiver/loadscraper`: Migrate the scraper to the mdatagen metrics builder (#7288)
 - `awsecscontainermetricsreceiver`: Rename attributes to follow semantic conventions (#7425)
-- `mysqlreceiver`: Add golden files for integration test
+- `datadogexporter`: Always map conventional attributes to tags (#7185)
+- `mysqlreceiver`: Add golden files for integration test (#7303)
+- `nginxreceiver`: Standardize integration test (#7515)
+- `mysqlreceiver`: Update to use mdatagen v2 (#7507)
+- `postgresqlreceiver`: Add integration tests (#7501)
+- `apachereceiver`: Add integration test (#7517)
+- `mysqlreceiver`: Use scrapererror to report errors (#7513)
+- `postgresreceiver`: Update to mdatagen v2 (#7503)
+- `nginxreceiver`: Update to mdatagen v2 (#7549)
+- `datadogexporter`: Fix traces exporter's initialization log (#7564)
+- `tailsamplingprocessor`: Add And sampling policy (#6910)
+- `coralogixexporter`: Add Coralogix Exporter (#7383)
+- `prometheusexecreceiver`: Add default value for `scrape_timeout` option (#7587)
 
 ## 🛑 Breaking changes 🛑
+
+- `resourcedetectionprocessor`: Update `os.type` attribute values according to semantic conventions (#7544)
+- `pkg/translator/prometheusremotewrite`: Cleanup prw translator public functions (#7776)
 
 ## 🚀 New components 🚀
 
@@ -23,6 +72,13 @@
 
 - `resourcedetectionprocessor`: fix `meta` allow list excluding keys with nil values (#7424)
 - `postgresqlreceiver`: Fix issue where empty metrics could be returned after failed connection (#7502)
+- `resourcetotelemetry`: Ensure resource attributes are added to summary
+  and exponential histogram data points. (#7523)
+
+## Deprecations
+
+- Deprecated otel_to_hec_fields.name setting from splunkhec exporter. (#7560)
+
 ## v0.43.0
 
 ## 💡 Enhancements 💡
@@ -35,7 +91,6 @@
 - `elasticsearchreceiver`: Implement scraping logic (#7174)
 - `datadogexporter`: Add http.status_code tag to trace stats (#6889)
 - `datadogexporter`: Add configuration option to use OTel span name into the Datatog resource name (#6611)
-- `dynatraceexporter`: Write error logs using plugin logger (#7360)
 - `mongodbreceiver`: Add initial client code to the component (#7125)
 - `tanzuobservabilityexporter`: Support delta histograms (#6897)
 - `awscloudwatchlogsexporter`: Use cwlogs package to export logs (#7152)